# app/backend/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DatasheetMiner API",
    "version": "1.0.0",
    "description": "Search and browse industrial product specifications extracted from datasheets. Covers motors, drives, gearheads, and robot arms. Free to use, no authentication required for read operations.",
    "contact": {
      "name": "DatasheetMiner",
      "url": "https://datasheets.advin.io"
    }
  },
  "servers": [
    {
      "url": "https://datasheets.advin.io",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/v1/search": {
      "get": {
        "summary": "Search products",
        "description": "Powerful search combining text matching, spec filtering, and multi-level sorting. Returns compact product summaries with key specs for each product type.",
        "operationId": "searchProducts",
        "tags": ["Search"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Text search query. Matches against part number, product name, manufacturer, series, product family, and type with weighted relevance scoring.",
            "schema": { "type": "string" },
            "example": "EC-45"
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by product type",
            "schema": { "type": "string", "enum": ["motor", "drive", "gearhead", "robot_arm"] }
          },
          {
            "name": "manufacturer",
            "in": "query",
            "description": "Filter by manufacturer (case-insensitive substring match)",
            "schema": { "type": "string" },
            "example": "Maxon"
          },
          {
            "name": "where",
            "in": "query",
            "description": "Spec filter expressions. Format: field{operator}value. Operators: >=, <=, !=, >, <, =. Numeric fields use numeric comparison; string fields use substring match. Repeat for multiple filters (AND logic).",
            "schema": { "type": "array", "items": { "type": "string" } },
            "style": "form",
            "explode": true,
            "examples": {
              "single": { "value": ["rated_power>=100"] },
              "multiple": { "value": ["rated_power>=100", "rated_voltage<=48"] }
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort keys. Format: field or field:desc. Repeat for multi-level sorting. Null values sort last.",
            "schema": { "type": "array", "items": { "type": "string" } },
            "style": "form",
            "explode": true,
            "examples": {
              "single": { "value": ["rated_power:desc"] },
              "multiple": { "value": ["rated_power:desc", "weight:asc"] }
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results (default 20, max 100)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ProductSummary" }
                    },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductSummary": {
        "type": "object",
        "description": "Compact product representation with identification fields and key specs for the product type. Returned by the search endpoint.",
        "properties": {
          "product_id": { "type": "string" },
          "product_type": { "type": "string" },
          "manufacturer": { "type": "string" },
          "product_name": { "type": "string" },
          "part_number": { "type": "string" },
          "relevance": { "type": "integer", "description": "Text search relevance score (0-100). Only present when a text query is provided." }
        },
        "additionalProperties": true,
        "required": ["product_id", "product_type", "manufacturer"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" }
        },
        "required": ["success", "error"]
      }
    }
  },
...
}